backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Max Dillon <maxdillon13124@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.3"
pydantic = "^2.9"
together = "^1.3"
python-dotenv = "^1.0"
vapi-server-sdk = "^0.1.0"
pypdf2 = "^3.0.1"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
django-cors-headers = "^4.6.0"
drf-spectacular = "^0.27.2"



[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pylint = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
